ProductVersie1/requirements.txt
aws-cdk-lib~=2.93.0
constructs>=10.0.0,<11.0.0
